refactor(aircraft-inject): move group classifier to mission_tools (decouple layers)
Address Sourcery overall feedback: convert_aircraft_groups depended on
aircrafts_injector (sibling-to-sibling). Move classify_aircraft_group + KIND_*
constants to the low-level mission_tools package, shared by both the injector
and the v5 converter. aircrafts_injector re-exports them for backward compat.
