+++ src/python/veaf-tools/aircrafts_injector/__init__.py
@@ -4,18 +4,23 @@ VEAF Aircrafts Injector Package
 This package provides classes for managing aircraft data and injecting aircraft groups into DCS missions.
 """
 
-from .aircraft_groups_extractor_README import AircraftGroupsExtractorREADME
-from .aircrafts_injector_injector_README import AircraftGroupsInjectorREADME
-from .aircrafts_injector_worker import (
+# Aircraft-group sort criteria live in the low-level mission_tools package
+# (shared with the v5 converter); re-exported here for backward compatibility.
+from mission_tools import (
     KIND_DYNAMIC_TEMPLATE,
     KIND_SPAWNABLE,
     SPAWNABLE_NAME_PREFIX,
+    classify_aircraft_group,
+)
+
+from .aircraft_groups_extractor_README import AircraftGroupsExtractorREADME
+from .aircrafts_injector_injector_README import AircraftGroupsInjectorREADME
+from .aircrafts_injector_worker import (
     AircraftGroupsExtractorWorker,
     AircraftGroupsInjectorWorker,
     AircraftGroupsYAMLValidator,
     InjectionResult,
     ValidationError,
-    classify_aircraft_group,
 )
 
 __all__ = [
